requirements.txt
tensorflow==2.14.0
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4
pillow==10.1.0
gunicorn==21.2.0
